Converted all input numerics to 64-bit to appease triton

+++ model_service/models/gpt2/model.py
@@ -44,15 +44,15 @@ class Model(AbstractModel):
                 Tensor(name="prompts", dtype=bytes, shape=(1,)),
                 Tensor(name='max_tokens', dtype=np.int64, shape=(1,), optional=True),
                 Tensor(name='min_tokens', dtype=np.int64, shape=(1,), optional=True),
-                Tensor(name='temperature', dtype=np.float32, shape=(1,), optional=True),
+                Tensor(name='temperature', dtype=np.float64, shape=(1,), optional=True),
                 Tensor(name='top_p', dtype=np.int64, shape=(1,), optional=True),
                 Tensor(name='top_k', dtype=np.int64, shape=(1,), optional=True),
-                Tensor(name='repetition_penalty', dtype=np.float32, shape=(1,), optional=True)
+                Tensor(name='repetition_penalty', dtype=np.float64, shape=(1,), optional=True)
             ],
             outputs=[
                 Tensor(name="sequences", dtype=object, shape=(-1,)),
                 Tensor(name="tokens", dtype=object, shape=(-1,)),
-                Tensor(name="logprobs", dtype=np.float32, shape=(-1,)),
+                Tensor(name="logprobs", dtype=np.float64, shape=(-1,)),
             ],
             config=ModelConfig(max_batch_size=128),
         )
